feat: move DTM saving before model creation
Can be helpfull if model creation fails, so that the DTM doesn't need to be recreated. If the current analysis is a replication and DTM is null, the saved DTM will be used. If instead one wants to used the saved one outside a replication, that parameter needs manual specification